MAINT: Move storage related modules to `map/_storage` (#396)
