+++ pipefunc/map/__init__.py
@@ -2,12 +2,17 @@
 
 from contextlib import suppress as _suppress
 
-from pipefunc.map._dictarray import DictArray, SharedMemoryDictArray
-from pipefunc.map._filearray import FileArray
 from pipefunc.map._mapspec import MapSpec
 from pipefunc.map._run import load_outputs, load_xarray_dataset, run
 from pipefunc.map._run_info import RunInfo
-from pipefunc.map._storage_base import StorageBase, register_storage, storage_registry
+from pipefunc.map._storage import (
+    DictArray,
+    FileArray,
+    SharedMemoryDictArray,
+    StorageBase,
+    register_storage,
+    storage_registry,
+)
 
 __all__ = [
     "DictArray",
@@ -24,6 +29,6 @@ __all__ = [
 ]
 
 with _suppress(ImportError):
-    from pipefunc.map.zarr import ZarrFileArray, ZarrMemoryArray, ZarrSharedMemoryArray
+    from pipefunc.map._storage import ZarrFileArray, ZarrMemoryArray, ZarrSharedMemoryArray
 
     __all__ += ["ZarrFileArray", "ZarrMemoryArray", "ZarrSharedMemoryArray"]
